+++ CHANGELOG.md
@@ -7,3 +7,11 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - Initial launch configurations and snippets are updated and fixed  
+
+## [0.0.8] - 2025-14-06  
+
+### Added
+
+- Added support for `Register` variables change  
+- Basic Linux syscalls `read (63)` and `write (64)` are now supported  
+- Basic terminal emulator support is added